drivers: escape*() methods moved to Engine [WIP]
