+++ src/Dibi/Drivers/Connection.php
@@ -74,24 +74,4 @@ interface Connection
 	function escapeText(string $value): string;
 
 	function escapeBinary(string $value): string;
-
-	function escapeIdentifier(string $value): string;
-
-	function escapeBool(bool $value): string;
-
-	function escapeDate(\DateTimeInterface $value): string;
-
-	function escapeDateTime(\DateTimeInterface $value): string;
-
-	function escapeDateInterval(\DateInterval $value): string;
-
-	/**
-	 * Encodes string for use in a LIKE statement.
-	 */
-	function escapeLike(string $value, int $pos): string;
-
-	/**
-	 * Injects LIMIT/OFFSET to the SQL query.
-	 */
-	function applyLimit(string &$sql, ?int $limit, ?int $offset): void;
 }
