+++ src/Dibi/Drivers/MySQLi/Connection.php
@@ -297,64 +297,4 @@ class Connection implements Drivers\Connection
 	{
 		return "_binary'" . $this->connection->escape_string($value) . "'";
 	}
-
-
-	public function escapeIdentifier(string $value): string
-	{
-		return '`' . str_replace('`', '``', $value) . '`';
-	}
-
-
-	public function escapeBool(bool $value): string
-	{
-		return $value ? '1' : '0';
-	}
-
-
-	public function escapeDate(\DateTimeInterface $value): string
-	{
-		return $value->format("'Y-m-d'");
-	}
-
-
-	public function escapeDateTime(\DateTimeInterface $value): string
-	{
-		return $value->format("'Y-m-d H:i:s.u'");
-	}
-
-
-	public function escapeDateInterval(\DateInterval $value): string
-	{
-		if ($value->y || $value->m || $value->d) {
-			throw new Dibi\NotSupportedException('Only time interval is supported.');
-		}
-
-		return $value->format("'%r%H:%I:%S.%f'");
-	}
-
-
-	/**
-	 * Encodes string for use in a LIKE statement.
-	 */
-	public function escapeLike(string $value, int $pos): string
-	{
-		$value = addcslashes(str_replace('\\', '\\\\', $value), "\x00\n\r\\'%_");
-		return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'");
-	}
-
-
-	/**
-	 * Injects LIMIT/OFFSET to the SQL query.
-	 */
-	public function applyLimit(string &$sql, ?int $limit, ?int $offset): void
-	{
-		if ($limit < 0 || $offset < 0) {
-			throw new Dibi\NotSupportedException('Negative offset or limit.');
-
-		} elseif ($limit !== null || $offset) {
-			// see http://dev.mysql.com/doc/refman/5.0/en/select.html
-			$sql .= ' LIMIT ' . ($limit ?? '18446744073709551615')
-				. ($offset ? ' OFFSET ' . $offset : '');
-		}
-	}
 }
