+++ src/Dibi/Drivers/PDO/Connection.php
@@ -220,165 +220,4 @@ class Connection implements Drivers\Connection
 			default => $this->connection->quote($value, PDO::PARAM_LOB),
 		};
 	}
-
-
-	public function escapeIdentifier(string $value): string
-	{
-		return match ($this->driverName) {
-			'mysql' => '`' . str_replace('`', '``', $value) . '`',
-			'oci', 'pgsql' => '"' . str_replace('"', '""', $value) . '"',
-			'sqlite' => '[' . strtr($value, '[]', '  ') . ']',
-			'odbc', 'mssql' => '[' . str_replace(['[', ']'], ['[[', ']]'], $value) . ']',
-			'dblib', 'sqlsrv' => '[' . str_replace(']', ']]', $value) . ']',
-			default => $value,
-		};
-	}
-
-
-	public function escapeBool(bool $value): string
-	{
-		if ($this->driverName === 'pgsql') {
-			return $value ? 'TRUE' : 'FALSE';
-		} else {
-			return $value ? '1' : '0';
-		}
-	}
-
-
-	public function escapeDate(\DateTimeInterface $value): string
-	{
-		return $value->format($this->driverName === 'odbc' ? '#m/d/Y#' : "'Y-m-d'");
-	}
-
-
-	public function escapeDateTime(\DateTimeInterface $value): string
-	{
-		return match ($this->driverName) {
-			'odbc' => $value->format('#m/d/Y H:i:s.u#'),
-			'mssql', 'dblib', 'sqlsrv' => 'CONVERT(DATETIME2(7), ' . $value->format("'Y-m-d H:i:s.u'") . ')',
-			default => $value->format("'Y-m-d H:i:s.u'"),
-		};
-	}
-
-
-	public function escapeDateInterval(\DateInterval $value): string
-	{
-		throw new Dibi\NotImplementedException;
-	}
-
-
-	/**
-	 * Encodes string for use in a LIKE statement.
-	 */
-	public function escapeLike(string $value, int $pos): string
-	{
-		switch ($this->driverName) {
-			case 'mysql':
-				$value = addcslashes(str_replace('\\', '\\\\', $value), "\x00\n\r\\'%_");
-				return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'");
-
-			case 'oci':
-				$value = addcslashes(str_replace('\\', '\\\\', $value), "\x00\\%_");
-				$value = str_replace("'", "''", $value);
-				return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'");
-
-			case 'pgsql':
-				$bs = substr($this->connection->quote('\\', PDO::PARAM_STR), 1, -1); // standard_conforming_strings = on/off
-				$value = substr($this->connection->quote($value, PDO::PARAM_STR), 1, -1);
-				$value = strtr($value, ['%' => $bs . '%', '_' => $bs . '_', '\\' => '\\\\']);
-				return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'");
-
-			case 'sqlite':
-				$value = addcslashes(substr($this->connection->quote($value, PDO::PARAM_STR), 1, -1), '%_\\');
-				return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'") . " ESCAPE '\\'";
-
-			case 'odbc':
-			case 'mssql':
-			case 'dblib':
-			case 'sqlsrv':
-				$value = strtr($value, ["'" => "''", '%' => '[%]', '_' => '[_]', '[' => '[[]']);
-				return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'");
-
-			default:
-				throw new Dibi\NotImplementedException;
-		}
-	}
-
-
-	/**
-	 * Injects LIMIT/OFFSET to the SQL query.
-	 */
-	public function applyLimit(string &$sql, ?int $limit, ?int $offset): void
-	{
-		if ($limit < 0 || $offset < 0) {
-			throw new Dibi\NotSupportedException('Negative offset or limit.');
-		}
-
-		switch ($this->driverName) {
-			case 'mysql':
-				if ($limit !== null || $offset) {
-					// see http://dev.mysql.com/doc/refman/5.0/en/select.html
-					$sql .= ' LIMIT ' . ($limit ?? '18446744073709551615')
-						. ($offset ? ' OFFSET ' . $offset : '');
-				}
-
-				break;
-
-			case 'pgsql':
-				if ($limit !== null) {
-					$sql .= ' LIMIT ' . $limit;
-				}
-
-				if ($offset) {
-					$sql .= ' OFFSET ' . $offset;
-				}
-
-				break;
-
-			case 'sqlite':
-				if ($limit !== null || $offset) {
-					$sql .= ' LIMIT ' . ($limit ?? '-1')
-						. ($offset ? ' OFFSET ' . $offset : '');
-				}
-
-				break;
-
-			case 'oci':
-				if ($offset) {
-					// see http://www.oracle.com/technology/oramag/oracle/06-sep/o56asktom.html
-					$sql = 'SELECT * FROM (SELECT t.*, ROWNUM AS "__rnum" FROM (' . $sql . ') t '
-						. ($limit !== null ? 'WHERE ROWNUM <= ' . ($offset + $limit) : '')
-						. ') WHERE "__rnum" > ' . $offset;
-
-				} elseif ($limit !== null) {
-					$sql = 'SELECT * FROM (' . $sql . ') WHERE ROWNUM <= ' . $limit;
-				}
-
-				break;
-
-			case 'mssql':
-			case 'sqlsrv':
-			case 'dblib':
-				// requires ORDER BY, see https://technet.microsoft.com/en-us/library/gg699618(v=sql.110).aspx
-				if ($limit !== null) {
-					$sql = sprintf('%s OFFSET %d ROWS FETCH NEXT %d ROWS ONLY', rtrim($sql), $offset, $limit);
-				} elseif ($offset) {
-					$sql = sprintf('%s OFFSET %d ROWS', rtrim($sql), $offset);
-				}
-
-				break;
-
-			case 'odbc':
-				if ($offset) {
-					throw new Dibi\NotSupportedException('Offset is not supported by this database.');
-
-				} elseif ($limit !== null) {
-					$sql = 'SELECT TOP ' . $limit . ' * FROM (' . $sql . ') t';
-					break;
-				}
-				// break omitted
-			default:
-				throw new Dibi\NotSupportedException('PDO or driver does not support applying limit or offset.');
-		}
-	}
 }
