+++ src/Dibi/Drivers/PgSQL/Connection.php
@@ -272,66 +272,4 @@ class Connection implements Drivers\Connection
 
 		return "'" . pg_escape_bytea($this->connection, $value) . "'";
 	}
-
-
-	public function escapeIdentifier(string $value): string
-	{
-		// @see http://www.postgresql.org/docs/8.2/static/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS
-		return '"' . str_replace('"', '""', $value) . '"';
-	}
-
-
-	public function escapeBool(bool $value): string
-	{
-		return $value ? 'TRUE' : 'FALSE';
-	}
-
-
-	public function escapeDate(\DateTimeInterface $value): string
-	{
-		return $value->format("'Y-m-d'");
-	}
-
-
-	public function escapeDateTime(\DateTimeInterface $value): string
-	{
-		return $value->format("'Y-m-d H:i:s.u'");
-	}
-
-
-	public function escapeDateInterval(\DateInterval $value): string
-	{
-		throw new Dibi\NotImplementedException;
-	}
-
-
-	/**
-	 * Encodes string for use in a LIKE statement.
-	 */
-	public function escapeLike(string $value, int $pos): string
-	{
-		$bs = pg_escape_string($this->connection, '\\'); // standard_conforming_strings = on/off
-		$value = pg_escape_string($this->connection, $value);
-		$value = strtr($value, ['%' => $bs . '%', '_' => $bs . '_', '\\' => '\\\\']);
-		return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'");
-	}
-
-
-	/**
-	 * Injects LIMIT/OFFSET to the SQL query.
-	 */
-	public function applyLimit(string &$sql, ?int $limit, ?int $offset): void
-	{
-		if ($limit < 0 || $offset < 0) {
-			throw new Dibi\NotSupportedException('Negative offset or limit.');
-		}
-
-		if ($limit !== null) {
-			$sql .= ' LIMIT ' . $limit;
-		}
-
-		if ($offset) {
-			$sql .= ' OFFSET ' . $offset;
-		}
-	}
 }
