+++ src/Dibi/Drivers/SQLSrv/Connection.php
@@ -204,63 +204,4 @@ class Connection implements Drivers\Connection
 	{
 		return '0x' . bin2hex($value);
 	}
-
-
-	public function escapeIdentifier(string $value): string
-	{
-		// @see https://msdn.microsoft.com/en-us/library/ms176027.aspx
-		return '[' . str_replace(']', ']]', $value) . ']';
-	}
-
-
-	public function escapeBool(bool $value): string
-	{
-		return $value ? '1' : '0';
-	}
-
-
-	public function escapeDate(\DateTimeInterface $value): string
-	{
-		return $value->format("'Y-m-d'");
-	}
-
-
-	public function escapeDateTime(\DateTimeInterface $value): string
-	{
-		return 'CONVERT(DATETIME2(7), ' . $value->format("'Y-m-d H:i:s.u'") . ')';
-	}
-
-
-	public function escapeDateInterval(\DateInterval $value): string
-	{
-		throw new Dibi\NotImplementedException;
-	}
-
-
-	/**
-	 * Encodes string for use in a LIKE statement.
-	 */
-	public function escapeLike(string $value, int $pos): string
-	{
-		$value = strtr($value, ["'" => "''", '%' => '[%]', '_' => '[_]', '[' => '[[]']);
-		return ($pos & 1 ? "'%" : "'") . $value . ($pos & 2 ? "%'" : "'");
-	}
-
-
-	/**
-	 * Injects LIMIT/OFFSET to the SQL query.
-	 */
-	public function applyLimit(string &$sql, ?int $limit, ?int $offset): void
-	{
-		if ($limit < 0 || $offset < 0) {
-			throw new Dibi\NotSupportedException('Negative offset or limit.');
-
-		} elseif ($limit !== null) {
-			// requires ORDER BY, see https://technet.microsoft.com/en-us/library/gg699618(v=sql.110).aspx
-			$sql = sprintf('%s OFFSET %d ROWS FETCH NEXT %d ROWS ONLY', rtrim($sql), $offset, $limit);
-		} elseif ($offset) {
-			// requires ORDER BY, see https://technet.microsoft.com/en-us/library/gg699618(v=sql.110).aspx
-			$sql = sprintf('%s OFFSET %d ROWS', rtrim($sql), $offset);
-		}
-	}
 }
