+++ src/Dibi/Translator.php
@@ -17,8 +17,9 @@ use function array_filter, array_keys, array_splice, array_values, count, explod
  */
 final class Translator
 {
-	private Connection $connection;
+	private readonly Connection $connection;
 	private readonly Drivers\Connection $driver;
+	private readonly Drivers\Engine $engine;
 	private int $cursor = 0;
 	private array $args;
 
@@ -36,6 +37,7 @@ final class Translator
 	{
 		$this->connection = $connection;
 		$this->driver = $connection->getDriver();
+		$this->engine = $connection->getDatabaseEngine();
 		$this->identifiers = new HashMap($this->delimite(...));
 	}
 
@@ -144,7 +146,7 @@ final class Translator
 
 		// apply limit
 		if ($this->limit !== null || $this->offset !== null) {
-			$this->driver->applyLimit($sql, $this->limit, $this->offset);
+			$this->engine->applyLimit($sql, $this->limit, $this->offset);
 		}
 
 		return $sql;
@@ -209,7 +211,7 @@ final class Translator
 				case 'n':  // key, key, ... identifier names
 					foreach ($value as $k => $v) {
 						if (is_string($k)) {
-							$vx[] = $this->identifiers->$k . (empty($v) ? '' : ' AS ' . $this->driver->escapeIdentifier($v));
+							$vx[] = $this->identifiers->$k . (empty($v) ? '' : ' AS ' . $this->engine->escapeIdentifier($v));
 						} else {
 							$pair = explode('%', $v, 2); // split into identifier & modifier
 							$vx[] = $this->identifiers->{$pair[0]};
@@ -346,7 +348,7 @@ final class Translator
 				case 's':  // string
 					return $value === null
 						? 'NULL'
-						: $this->driver->escapeText((string) $value);
+						: $this->engine->escapeText((string) $value);
 
 				case 'bin':// binary
 					return $value === null
@@ -356,7 +358,7 @@ final class Translator
 				case 'b':  // boolean
 					return $value === null
 						? 'NULL'
-						: $this->driver->escapeBool((bool) $value);
+						: $this->engine->escapeBool((bool) $value);
 
 				case 'sN': // string or null
 				case 'sn':
@@ -406,15 +408,15 @@ final class Translator
 					}
 
 					return $modifier === 'd'
-						? $this->driver->escapeDate($value)
-						: $this->driver->escapeDateTime($value);
+						? $this->engine->escapeDate($value)
+						: $this->engine->escapeDateTime($value);
 
 				case 'by':
 				case 'n':  // composed identifier name
 					return $this->identifiers->$value;
 
 				case 'N':  // identifier name
-					return $this->driver->escapeIdentifier($value);
+					return $this->engine->escapeIdentifier($value);
 
 				case 'ex':
 				case 'sql': // preserve as dibi-SQL  (TODO: leave only %ex)
@@ -450,16 +452,16 @@ final class Translator
 					return (string) $value;
 
 				case 'like~':  // LIKE string%
-					return $this->driver->escapeLike($value, 2);
+					return $this->engine->escapeLike($value, 2);
 
 				case '~like':  // LIKE %string
-					return $this->driver->escapeLike($value, 1);
+					return $this->engine->escapeLike($value, 1);
 
 				case '~like~': // LIKE %string%
-					return $this->driver->escapeLike($value, 3);
+					return $this->engine->escapeLike($value, 3);
 
 				case 'like': // LIKE string
-					return $this->driver->escapeLike($value, 0);
+					return $this->engine->escapeLike($value, 0);
 
 				case 'and':
 				case 'or':
@@ -485,16 +487,16 @@ final class Translator
 			return rtrim(rtrim(number_format($value, 10, '.', ''), '0'), '.');
 
 		} elseif (is_bool($value)) {
-			return $this->driver->escapeBool($value);
+			return $this->engine->escapeBool($value);
 
 		} elseif ($value === null) {
 			return 'NULL';
 
 		} elseif ($value instanceof \DateTimeInterface) {
-			return $this->driver->escapeDateTime($value);
+			return $this->engine->escapeDateTime($value);
 
 		} elseif ($value instanceof \DateInterval) {
-			return $this->driver->escapeDateInterval($value);
+			return $this->engine->escapeDateInterval($value);
 
 		} elseif ($value instanceof Literal) {
 			return (string) $value;
@@ -653,7 +655,7 @@ final class Translator
 		$parts = explode('.', $value);
 		foreach ($parts as &$v) {
 			if ($v !== '*') {
-				$v = $this->driver->escapeIdentifier($v);
+				$v = $this->engine->escapeIdentifier($v);
 			}
 		}
