+++ tests/dibi/Connection.objectTranslator.phpt
@@ -49,7 +49,7 @@ test('DateTime', function () use ($conn) {
 
 	// Without object translator, DateTime child is translated by driver
 	Assert::same(
-		$conn->getDriver()->escapeDateTime($stamp),
+		$conn->getDatabaseEngine()->escapeDateTime($stamp),
 		$conn->translate('?', $stamp),
 	);
 
@@ -67,15 +67,15 @@ test('DateTime', function () use ($conn) {
 
 	// With modifier, it is still translated by driver
 	Assert::same(
-		$conn->getDriver()->escapeDateTime($stamp),
+		$conn->getDatabaseEngine()->escapeDateTime($stamp),
 		$conn->translate('%dt', $stamp),
 	);
 	Assert::same(
-		$conn->getDriver()->escapeDateTime($stamp),
+		$conn->getDatabaseEngine()->escapeDateTime($stamp),
 		$conn->translate('%t', $stamp),
 	);
 	Assert::same(
-		$conn->getDriver()->escapeDate($stamp),
+		$conn->getDatabaseEngine()->escapeDate($stamp),
 		$conn->translate('%d', $stamp),
 	);
 
@@ -83,7 +83,7 @@ test('DateTime', function () use ($conn) {
 	// DateTimeImmutable as a Time parent is not affected and still translated by driver
 	$dt = DateTimeImmutable::createFromFormat('Y-m-d H:i:s', '2022-11-22 12:13:14');
 	Assert::same(
-		$conn->getDriver()->escapeDateTime($dt),
+		$conn->getDatabaseEngine()->escapeDateTime($dt),
 		$conn->translate('?', $dt),
 	);
